// src/types.ts
export enum Currencies {
  RUB = 'RUB', // ✅
  BYN = 'BYN', // ✅
  USD = 'USD', // ✅
  EUR = 'EUR', // ✅
  PLN = 'PLN', // ✅
  CNY = 'CNY', // ✅
  AUD = 'AUD', // ✅
  AZN = 'AZN', // ✅
  GBP = 'GBP', // ✅
  AMD = 'AMD', // ✅
  BGN = 'BGN', // ✅
  BRL = 'BRL', // ✅
  HUF = 'HUF', // ✅
  VND = 'VND', // ✅
  HKD = 'HKD', // ✅
  GEL = 'GEL', // ✅
  DKK = 'DKK', // ✅
  AED = 'AED', // ✅
  EGP = 'EGP', // ✅
  INR = 'INR', // ✅
  IDR = 'IDR', // ✅
  KZT = 'KZT', // ✅
  CAD = 'CAD', // ✅
  QAR = 'QAR', // ✅
  KGS = 'KGS', // ✅
  MDL = 'MDL', // ✅
  NZD = 'NZD', // ✅
  NOK = 'NOK', // ✅
  RON = 'RON', // ✅
  XDR = 'XDR', // ✅
  SGD = 'SGD', // ✅
  TJS = 'TJS', // ✅
  THB = 'THB', // ✅
  TRY = 'TRY', // ✅
  TMT = 'TMT', // ✅
  UZS = 'UZS', // ✅
  UAH = 'UAH', // ✅
  CZK = 'CZK', // ✅
  SEK = 'SEK', // ✅
  CHF = 'CHF', // ✅
  RSD = 'RSD', // ✅
  ZAR = 'ZAR', // ✅
  KRW = 'KRW', // ✅
  JPY = 'JPY',
}